beginner_level/s3_lifeCycle/requirements.txt
-e .
aws.cdk.aws_s3
